+++ Code/src/GUI/EquipmentGUI/ItemInfoPanel.java
@@ -55,8 +55,8 @@ public class ItemInfoPanel extends JPanel {
     }
 
     private class DiceItemInfoPanel extends JPanel{
-        private final static double MANA_COST_TO_ALL_PROPORTION=0.3;
-        private JLabel nameLabel, requirements, manaCost;
+        private final static double MANA_COST_TO_ALL_PROPORTION=0.5;
+        private JLabel nameLabel, requirements, manaCost, target,attribute;
         private DiceSidesPanel diceSidesPanel;
 
         public DiceItemInfoPanel() {
@@ -66,36 +66,50 @@ public class ItemInfoPanel extends JPanel {
             nameLabel = new JLabel("", SwingConstants.CENTER);
             nameLabel.setFont(GUISettings.BIG_FONT);
             nameLabel.setForeground(Color.WHITE);
-            nameLabel.setPreferredSize(new Dimension(GUISettings.PANEL_SIZE-10,GUISettings.SMALL_PANEL_SIZE/5));
+            nameLabel.setPreferredSize(new Dimension(GUISettings.PANEL_SIZE-10,GUISettings.SMALL_PANEL_SIZE/7));
             nameLabel.setVerticalAlignment(SwingConstants.CENTER);
 
             diceSidesPanel=new DiceSidesPanel();
 
             requirements = new JLabel();
             requirements.setForeground(Color.WHITE);
-            requirements.setPreferredSize(new Dimension((int)((GUISettings.PANEL_SIZE)*(1-MANA_COST_TO_ALL_PROPORTION)-20),GUISettings.SMALL_PANEL_SIZE/10));
+            requirements.setPreferredSize(new Dimension((int)((GUISettings.PANEL_SIZE)*MANA_COST_TO_ALL_PROPORTION-20),GUISettings.SMALL_PANEL_SIZE/10));
+
+            target = new JLabel();
+            target.setForeground(Color.WHITE);
+            target.setPreferredSize(new Dimension((int)((GUISettings.PANEL_SIZE)*MANA_COST_TO_ALL_PROPORTION-20),GUISettings.SMALL_PANEL_SIZE/10));
+
+            attribute = new JLabel();
+            attribute.setForeground(Color.WHITE);
+            attribute.setPreferredSize(new Dimension((int)((GUISettings.PANEL_SIZE)*MANA_COST_TO_ALL_PROPORTION-20),GUISettings.SMALL_PANEL_SIZE/10));
 
             manaCost = new JLabel();
             manaCost.setForeground(Color.WHITE);
-            manaCost.setPreferredSize(new Dimension((int)((GUISettings.PANEL_SIZE-10)*MANA_COST_TO_ALL_PROPORTION),GUISettings.SMALL_PANEL_SIZE/10));
+            manaCost.setPreferredSize(new Dimension((int)((GUISettings.PANEL_SIZE)*MANA_COST_TO_ALL_PROPORTION-20),GUISettings.SMALL_PANEL_SIZE/10));
 
             this.add(nameLabel);
             this.add(diceSidesPanel);
             this.add(requirements);
+            this.add(target);
+            this.add(attribute);
             this.add(manaCost);
         }
 
         public void refresh(){
             Item item = GameManager.getEquipment().getPointedItem();
 
+            nameLabel.setText(item.name);
+
             if(item instanceof ActionItem aItem) {
-                nameLabel.setText(aItem.name);
                 diceSidesPanel.setDiceSides(aItem.getAction().getDice().getSides());
                 manaCost.setText("");
+                target.setText("Target: "+aItem.getAction().getTarget().toString());
+                attribute.setText("Amplify by: "+aItem.getScaleAttribute());
             } else if (item instanceof SpellItem sItem) {
-                nameLabel.setText(sItem.name);
                 diceSidesPanel.setDiceSides(sItem.getAction().getDice().getSides());
                 manaCost.setText("Mana: "+sItem.getAction().getManaCost());
+                target.setText("target: "+sItem.getAction().getTarget().toString());
+                attribute.setText("amplify by: "+sItem.getScaleAttribute());
             }
 
             StringBuilder requirementsBuilder=new StringBuilder("Requirements:");
@@ -183,7 +197,7 @@ public class ItemInfoPanel extends JPanel {
 
     private class DiceLessItemInfoPanel extends JPanel{
 
-        private JLabel nameLabel, effectLabel,quantityLabel;
+        private JLabel nameLabel, effectLabel,quantityLabel, targetLabel;
 
         public DiceLessItemInfoPanel() {
             this.setSize(GUISettings.PANEL_SIZE,GUISettings.SMALL_PANEL_SIZE);
@@ -199,17 +213,23 @@ public class ItemInfoPanel extends JPanel {
             effectLabel.setPreferredSize(new Dimension(GUISettings.PANEL_SIZE-20,GUISettings.SMALL_PANEL_SIZE/7));
             effectLabel.setForeground(Color.WHITE);
 
+            targetLabel = new JLabel("", SwingConstants.LEFT);
+            targetLabel.setPreferredSize(new Dimension(GUISettings.PANEL_SIZE-20,GUISettings.SMALL_PANEL_SIZE/7));
+            targetLabel.setForeground(Color.WHITE);
+
             quantityLabel = new JLabel("",SwingConstants.LEFT);
             quantityLabel.setPreferredSize(new Dimension(GUISettings.PANEL_SIZE-20,GUISettings.SMALL_PANEL_SIZE/7));
             quantityLabel.setForeground(Color.WHITE);
 
             this.add(nameLabel);
             this.add(effectLabel);
+            this.add(targetLabel);
             this.add(quantityLabel);
         }
 
         void refresh(){
             UsableItem item = (UsableItem) GameManager.getEquipment().getPointedItem();
+            targetLabel.setText("Target: "+item.getAction().getTarget().toString());
             StringBuilder builder=new StringBuilder("Effects:");
 
             var x = item.getAction().getActionFactories();
@@ -234,12 +254,13 @@ public class ItemInfoPanel extends JPanel {
         private JLabel[] diceSides;
 
         public DiceSidesPanel() {
-            this.setSize(GUISettings.PANEL_SIZE-10,GUISettings.SMALL_PANEL_SIZE/2);
+            int diceSize= Math.min((GUISettings.PANEL_SIZE-10)/6,GUISettings.SMALL_PANEL_SIZE/2);
+            this.setPreferredSize(new Dimension(diceSize*6+6,diceSize));
             this.setBackground(Color.BLACK);
             FlowLayout diceLayout = new FlowLayout(FlowLayout.CENTER);
             diceLayout.setHgap(1);
             this.setLayout(diceLayout);
-            int diceSize= Math.min((GUISettings.PANEL_SIZE-10)/6,GUISettings.SMALL_PANEL_SIZE/2);
+
             this.diceSides = new JLabel[6];
 
             for(int i=0;i<6;i++){