+++ changelog.md
@@ -3,6 +3,8 @@
 ### v0.2
 - Change status placeholders
 - Fix enemy shield bug
+- Fix GUI minor issue
+- Add target and attribute info to items
 ### v0.1
 - Game is playable and can be finished
 - It's first version that was distributed